split add field input to separate component
